move wallet providers to it's own directory
