+++ apps/web/core/test/core.test.ts
@@ -5,10 +5,11 @@ import { Poll } from "chain/dist/runtime/modules/poll";
 import { UInt64 } from "@proto-kit/library";
 import { PollManagerController } from "../controllers/poll-manager-controller";
 import { WalletController } from "../controllers/wallet-controller";
-import { WalletProvider } from "../providers/wallet-provider";
+import { WalletProvider } from "../providers/wallets/wallet-provider";
 import { PrivateKey } from "o1js";
 import { InMemoryPollStore } from "../controllers/poll-store";
 import { ChainTestController } from "./test-utils/chain-test-controller";
+import { PollController } from "../controllers/poll-controller";
 
 const PRIVATE_KEY = "EKDii5d1dA7DDw6NZwN7jF7qcdYR5MVjZ9TfESv1gc2TvmvV2WAE";
 
@@ -57,7 +58,7 @@ describe("Poll Manager", () => {
       store,
     });
 
-    const pollId = await pollManager.create({
+    const pollResult = await pollManager.create({
       title: "Test poll",
       description: "Test poll description",
       options: ["Option 1", "Option 2"],
@@ -65,6 +66,22 @@ describe("Poll Manager", () => {
       salt: "salt",
     });
 
-    expect(pollId).toBeDefined();
+    expect(pollResult).toBeDefined();
+
+    expect(pollResult.id).toBe(1);
+
+    const poll = await new PollController({
+      client: appChain,
+      wallet,
+      chain,
+    });
+
+    await poll.loadPoll(pollResult.id);
+
+    expect(poll.votes.length).toBe(10);
+
+    expect(poll.votes[0].votesCount).toBe(0);
+
+    await poll.vote(pollResult.id, poll.votes[0].hash);
   });
 });
